Split module into multiple files
We're going to be handling OpenAPI 2.0 and 3.0.0 in different modules
shortly. Set everything up for this.
